refactor: split main.py routes into routes/ package
routes/decks.py   — deck + preset endpoints
routes/review.py  — today, today-mixed, review, bury/unbury
routes/story.py   — story, regenerate, speak, preload, preload-session
routes/browse.py  — browse, stats, word, import
routes/utils.py   — shared DISABLE_AI flag and leaf_ids helper

main.py is now only app setup, lifespan, router registration, and CLI.
